+++ kmath-core/src/commonMain/kotlin/space/kscience/kmath/linear/MatrixBuilder.kt
@@ -5,7 +5,7 @@
 
 package space.kscience.kmath.linear
 
-import space.kscience.attributes.FlagAttribute
+import space.kscience.attributes.Attributes
 import space.kscience.attributes.SafeType
 import space.kscience.attributes.WithType
 import space.kscience.kmath.UnstableKMathAPI
@@ -13,62 +13,51 @@ import space.kscience.kmath.operations.Ring
 import space.kscience.kmath.structures.BufferAccessor2D
 import space.kscience.kmath.structures.MutableBufferFactory
 
-public class MatrixBuilder<T : Any, out A : Ring<T>>(
+/**
+ * A builder for matrix with fixed size
+ */
+@UnstableKMathAPI
+public class MatrixBuilder<T, out A : Ring<T>>(
     public val linearSpace: LinearSpace<T, A>,
-    public val rows: Int,
-    public val columns: Int,
+    public val rowNum: Int,
+    public val colNum: Int,
 ) : WithType<T> {
 
     override val type: SafeType<T> get() = linearSpace.type
 
-    public operator fun invoke(vararg elements: T): Matrix<T> {
-        require(rows * columns == elements.size) { "The number of elements ${elements.size} is not equal $rows * $columns" }
-        return linearSpace.buildMatrix(rows, columns) { i, j -> elements[i * columns + j] }
-    }
+}
 
-    //TODO add specific matrix builder functions like diagonal, etc
+@UnstableKMathAPI
+public fun <T, A : Ring<T>> MatrixBuilder<T, A>.sparse(): SparseMatrix<T> =
+    SparseMatrix(rowNum, colNum, linearSpace.elementAlgebra.zero)
+
+@UnstableKMathAPI
+public fun <T, A : Ring<T>> MatrixBuilder<T, A>.fill(vararg elements: T): Matrix<T> {
+    require(rowNum * colNum == elements.size) { "The number of elements ${elements.size} is not equal $rowNum * $colNum" }
+    return linearSpace.buildMatrix(rowNum, colNum) { i, j -> elements[i * colNum + j] }
 }
 
 /**
  * Create a matrix builder with given number of rows and columns
  */
 @UnstableKMathAPI
-public fun <T : Any, A : Ring<T>> LinearSpace<T, A>.matrix(rows: Int, columns: Int): MatrixBuilder<T, A> =
+public fun <T, A : Ring<T>> LinearSpace<T, A>.MatrixBuilder(rows: Int, columns: Int): MatrixBuilder<T, A> =
     MatrixBuilder(this, rows, columns)
 
-@UnstableKMathAPI
-public fun <T : Any> LinearSpace<T, Ring<T>>.vector(vararg elements: T): Point<T> {
-    return buildVector(elements.size) { elements[it] }
-}
-
-public inline fun <T : Any> LinearSpace<T, Ring<T>>.row(
-    size: Int,
-    crossinline builder: (Int) -> T,
-): Matrix<T> = buildMatrix(1, size) { _, j -> builder(j) }
-
-public fun <T : Any> LinearSpace<T, Ring<T>>.row(vararg values: T): Matrix<T> = row(values.size, values::get)
-
-public inline fun <T : Any> LinearSpace<T, Ring<T>>.column(
-    size: Int,
-    crossinline builder: (Int) -> T,
-): Matrix<T> = buildMatrix(size, 1) { i, _ -> builder(i) }
-
-public fun <T : Any> LinearSpace<T, Ring<T>>.column(vararg values: T): Matrix<T> = column(values.size, values::get)
-
-public object Symmetric : MatrixAttribute<Unit>, FlagAttribute
-
 /**
- * Naive implementation of a symmetric matrix builder, that adds a [Symmetric] tag. The resulting matrix contains
- * full `size^2` number of elements, but caches elements during calls to save [builder] calls. [builder] is always called in the
- * upper triangle region meaning that `i <= j`
+ * Naive implementation of a symmetric matrix builder, that adds a [Symmetric] tag.
+ * The resulting matrix contains full `size^2` number of elements,
+ * but caches elements during calls to save [builder] calls.
+ * Always called in the upper triangle region meaning that `i <= j`
  */
-public fun <T : Any, A : Ring<T>> MatrixBuilder<T, A>.symmetric(
-    builder: (i: Int, j: Int) -> T,
+@UnstableKMathAPI
+public fun <T, A : Ring<T>> MatrixBuilder<T, A>.symmetric(
+    builder: A.(i: Int, j: Int) -> T,
 ): Matrix<T> {
-    require(columns == rows) { "In order to build symmetric matrix, number of rows $rows should be equal to number of columns $columns" }
-    return with(BufferAccessor2D<T?>(rows, rows, MutableBufferFactory(type))) {
+    require(colNum == rowNum) { "In order to build symmetric matrix, number of rows $rowNum should be equal to number of columns $colNum" }
+    return with(BufferAccessor2D<T?>(rowNum, rowNum, MutableBufferFactory(type))) {
         val cache = HashMap<IntArray, T>()
-        linearSpace.buildMatrix(rows, rows) { i, j ->
+        linearSpace.buildMatrix(this@symmetric.rowNum, this@symmetric.rowNum) { i, j ->
             val index = intArrayOf(i, j)
             val cached = cache[index]
             if (cached == null) {
@@ -82,3 +71,49 @@ public fun <T : Any, A : Ring<T>> MatrixBuilder<T, A>.symmetric(
         }.withAttribute(Symmetric)
     }
 }
+
+/**
+ * Create a diagonal matrix with given factory.
+ */
+@UnstableKMathAPI
+public fun <T, A : Ring<T>> MatrixBuilder<T, A>.diagonal(
+    builder: A.(Int) -> T
+): Matrix<T> = with(linearSpace.elementAlgebra) {
+    require(colNum == rowNum) { "In order to build symmetric matrix, number of rows $rowNum should be equal to number of columns $colNum" }
+    return VirtualMatrix(rowNum, colNum, attributes = Attributes(IsDiagonal)) { i, j ->
+        check(i in 0 until rowNum) { "$i out of bounds: 0..<$rowNum" }
+        check(j in 0 until colNum) { "$j out of bounds: 0..<$colNum" }
+        if (i == j) {
+            builder(i)
+        } else {
+            zero
+        }
+    }
+}
+
+/**
+ * Create a diagonal matrix from elements
+ */
+@UnstableKMathAPI
+public fun <T> MatrixBuilder<T, Ring<T>>.diagonal(vararg elements: T): Matrix<T> {
+    require(colNum == rowNum) { "In order to build symmetric matrix, number of rows $rowNum should be equal to number of columns $colNum" }
+    return return VirtualMatrix(rowNum, colNum, attributes = Attributes(IsDiagonal)) { i, j ->
+        check(i in 0 until rowNum) { "$i out of bounds: 0..<$rowNum" }
+        check(j in 0 until colNum) { "$j out of bounds: 0..<$colNum" }
+        if (i == j) {
+            elements[i]
+        } else {
+            linearSpace.elementAlgebra.zero
+        }
+    }
+}
+
+/**
+ * Create a lazily evaluated virtual matrix with a given size
+ */
+@UnstableKMathAPI
+public fun <T : Any> MatrixBuilder<T, *>.virtual(
+    attributes: Attributes = Attributes.EMPTY,
+    generator: (i: Int, j: Int) -> T,
+): VirtualMatrix<T> = VirtualMatrix(rowNum, colNum, attributes, generator)
+